schema version rejection
