+++ editor/package.json
@@ -163,6 +163,7 @@
     "react-window": "^1.8.6",
     "recharts": "^2.12.7",
     "resend": "^3.2.0",
+    "semver": "^7.7.1",
     "signature_pad": "^4.2.0",
     "stylis": "^4.3.2",
     "svg-pathdata": "^7.1.0",
@@ -201,6 +202,7 @@
     "@types/react": "^18",
     "@types/react-dom": "^18",
     "@types/react-window": "^1.8.8",
+    "@types/semver": "^7.5.8",
     "@types/stylis": "^4.2.6",
     "@types/ua-parser-js": "^0.7.39",
     "@types/uuid": "^9.0.8",
