+++ editor/scaffolds/editor/state.ts
@@ -98,10 +98,13 @@ export interface FormDocumentEditorInit extends BaseDocumentEditorInit {
   form_id: string;
   campaign: EditorState["form"]["campaign"];
   form_security: EditorState["form"]["form_security"];
-  start:
-    | // TODO: review me - only need single scene
-    (grida.program.document.Document & { template_id: string })
-    | null;
+
+  /**
+   * the start document as-is (the typing is ignored - this should be assured before being actually passed under the provider)
+   *
+   * TODO: review me - only need single scene
+   */
+  start: FormStartPageSchema | unknown;
   ending: EditorState["form"]["ending"];
   connections?: {
     store_id?: number | null;
@@ -389,6 +392,30 @@ interface IEditorPagesState {
   selected_page_id?: "form" | "form/startpage" | "site" | (string & {});
 }
 
+export type SchemaMayVaryDocument<S> = {
+  /**
+   * the schema version from the server
+   */
+  __schema_version: string;
+} & (
+  | {
+      /**
+       * if the schema is validated (the version is correct)
+       */
+      __schema_valid: true;
+
+      state: S;
+    }
+  | {
+      /**
+       * if the schema is validated (the version is correct)
+       */
+      __schema_valid: false;
+
+      state: null;
+    }
+);
+
 export interface BaseDocumentEditorState
   extends IEditorGlobalSavingState,
     IEditorDateContextState,
@@ -413,9 +440,11 @@ export interface BaseDocumentEditorState
   document_title: string;
   doctype: GDocumentType;
   documents: {
-    ["site"]?: IDocumentEditorState;
-    ["canvas"]?: IDocumentEditorState;
-    ["form/startpage"]?: IDocumentEditorState & { template_id: string };
+    ["site"]?: SchemaMayVaryDocument<IDocumentEditorState>;
+    ["canvas"]?: SchemaMayVaryDocument<IDocumentEditorState>;
+    ["form/startpage"]?:
+      | SchemaMayVaryDocument<IDocumentEditorState & { template_id: string }>
+      | undefined;
     // [key: string]: ITemplateEditorState;
   };
   theme: {
